--- TypTop.Game/App.xaml
@@ -1,9 +1,0 @@
-﻿<Application 
-    x:Class="TypTop.Game.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    StartupUri="MainWindow.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
