--- TypTop.Game/App.xaml.cs
@@ -1,8 +1,0 @@
-﻿using System.Windows;
-
-namespace TypTop.Game
-{
-    public partial class App : Application
-    {
-    }
-}
