--- TypTop.Game/MainWindow.xaml
@@ -1,44 +1,0 @@
-﻿<Window 
-    x:Class="TypTop.Game.MainWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:monoGameControls="clr-namespace:TypTop.Game.MonoGameControls"
-    xmlns:local="clr-namespace:TypTop.Game"
-    mc:Ignorable="d"
-    Title="TypTop.Game" 
-    WindowStartupLocation="CenterScreen"
-    Width="800"
-    Height="480"
-    TextInput="MainWindow_OnTextInput"
-    >
-    <DockPanel LastChildFill="True">
-        <Menu DockPanel.Dock="Top">
-            <MenuItem Header="_File">
-                <MenuItem Header="_New" />
-                <MenuItem Header="_Open..." />
-                <Separator />
-                <MenuItem Header="_Save" />
-                <MenuItem Header="Save _As..." />
-                <Separator />
-                <MenuItem Header="E_xit" />
-            </MenuItem>
-            <MenuItem Header="_Edit">
-                <MenuItem Header="_Undo" />
-                <MenuItem Header="_Redo" />
-                <Separator />
-                <MenuItem Header="Cu_t" />
-                <MenuItem Header="_Copy" />
-                <MenuItem Header="_Paste" />
-                <MenuItem Header="_Delete" />
-                <Separator />
-                <MenuItem Header="Select _All" />
-            </MenuItem>
-            <MenuItem Header="_Help">
-                <MenuItem Header="_About TypTop.Game" />
-            </MenuItem>
-        </Menu>
-        <monoGameControls:MonoGameContentControl />
-    </DockPanel>
-</Window>
