--- TypTop.Game/ViewModel.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel;
-using System.Runtime.CompilerServices;
-
-namespace TypTop.Game
-{
-    public class ViewModel : INotifyPropertyChanged
-    {
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-    }
-}
