+++ TypTop.Logic/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace TypTop.Logic
+{
+    public class Class1
+    {
+    }
+}
