+++ TypTop.VisualKeyboard/TypTop.VisualKeyboard.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <UseWPF>true</UseWPF>
+    <RootNamespace>TypTop</RootNamespace>
+  </PropertyGroup>
+
+</Project>
